requirements.txt
torch==1.8.1
torchvision==0.9.1
pytorch_lightning==1.3.5
torch_optimizer==0.1.0
numpy==1.20.3
imageio==2.9.0
Pillow==8.2.0
matplotlib==3.4.2
fire==0.4.0
